pyproject.toml
[project]
name = "decimatr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
]

[dependency-groups]
cpu-only = [
    "decord[cpu]>=0.6.0",
    "imagehash>=4.3.2",
    "pytest>=8.3.5",
]
gpu = [
    "open-clip-torch>=2.32.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
